Refactor imports for consistency; add unit tests for predictors and enhance NGramUtil setup
