+++ convassist/tests/predictors/test_general_word_predictor.py
@@ -14,8 +14,8 @@ from convassist.predictor.smoothed_ngram_predictor.general_word_predictor import
 )
 from convassist.utilities.ngram.ngramutil import NGramUtil
 
-from convassist import setup_utils
-from convassist import TestPredictors
+from convassist.tests import setup_utils
+from convassist.tests.predictors import TestPredictors
 
 
 class TestGeneralWordPredictor(TestPredictors):
